Add in conditions metadata
Add in metadata for Random number seed provider and ecal hex readout.
The actual detector version used is passed to metadata extraction
such that the right conditions are parsed.

NOTE that this won't work for v10 or v11 but implementing this seems
like overkill for now, they are historical only.