Add method to copy input files from local storage to working directory
Since ARC can't copy local data, and the singularity image can't see
any (unmounted) other directories, input has to be brought to the working
directory on the node. This commit adds a command line argument `copy-local`
that is always called by ldmxsim,sh, but the corresponding function in the
helper script only attempts to copy a file if the job config lists a local
RSE that is used for input.