Adding functionality to retain crucial input meta instead of overwrite
For now this is any inputfile listed, and the LdmxImage (the singularity
image used to generate the input file). These need to be there for
posterity when there are multiple reconstruction passes run on a
chain of input files, i.e. when reconstruction run on something that was
already using input.

An example is trigger skimming, or signal reconstruction (where the gen
madgraph image is different from the ldmx-sw image used in simulation
and potentially also different again later in reconstruction).