Tweak to make metadata extraction work with singularity; give up
Tried different ways of loading the json from singularity inspect
Realized that the problem is that older versions of singularity
don't read in whatever would be the output of docker inspect
So instead adapted to skipping trying, if there is no metadata output
shipped with the command. Once this works, we can invoke that part
of the function from the command line.